+++ LoadM3d.h
@@ -44,17 +44,17 @@ public:
 	/*
 	bool LoadM3d(const std::string& filename, 
 		std::vector<PosNormalTex>& vertices,
-		std::vector<unsigned short>& indices,
+		std::vector<unsigned int>& indices,
 		std::vector<Subset>& subsets,
 		std::vector<M3dMaterial>& mats);
 	bool LoadM3d(const std::string& filename, 
 		std::vector<PosNormalTexTan>& vertices,
-		std::vector<unsigned short>& indices,
+		std::vector<unsigned int>& indices,
 		std::vector<Subset>& subsets,
 		std::vector<M3dMaterial>& mats);*/
 	bool LoadM3d(const std::string& filename, 
 		void*& vertices,
-		std::vector<unsigned short>& indices,
+		std::vector<unsigned int>& indices,
 		std::vector<Subset>& subsets,
 		std::vector<M3dMaterial>& mats,
 		AnimationData& skinInfo, std::string vertextype, size_t& numVertex);
@@ -66,7 +66,7 @@ private:
 	//void ReadVertices(std::ifstream& fin, UINT numVertices, std::vector<PosNormalTexTan>& vertices);
 	//void ReadSkinnedVertices(std::ifstream& fin, UINT numVertices, std::vector<PosNormalTexTanAnimation>& vertices);
 	void ReadSkinnedVertices(std::ifstream& fin, UINT numVertices, void*& vertices, std::string vertextype);
-	void ReadTriangles(std::ifstream& fin, UINT numTriangles, std::vector<unsigned short>& indices);
+	void ReadTriangles(std::ifstream& fin, UINT numTriangles, std::vector<unsigned int>& indices);
 	void ReadBoneOffsets(std::ifstream& fin, UINT numBones, std::vector<XMFLOAT4X4>& boneOffsets);
 	void ReadBoneHierarchy(std::ifstream& fin, UINT numBones, std::vector<int>& boneIndexToParentIndex);
 	void ReadAnimationClips(std::ifstream& fin, UINT numBones, UINT numAnimationClips, std::unordered_map<std::string, AnimationClip>& animations);
